apps/backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Data Collection
psutil>=5.9.8
httpx>=0.26.0
mss>=9.0.1
pillow>=10.2.0

# AI
openai>=1.10.0

# PDF
reportlab>=4.0.0

# Security
keyring>=24.3.0
cryptography>=41.0.0

# Utilities
websockets>=12.0
apscheduler>=3.10.4
aiofiles>=23.2.1

# Stripe (optional)
stripe>=7.0.0

# Cloud Storage & Database
firebase-admin>=6.2.0
supabase>=2.0.0

# ActivityWatch client (optional for desktop)
# aw-client>=0.5.13
